pyproject.toml
[build-system]
requires = ["hatchling>=1.8.0", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "NeuralProcesses"
description = "A framework for composing Neural Processes in Python"
authors = [
    {name="Wessel Bruinsma", email="wessel.p.bruinsma@gmail.com"},
]
license = {text="MIT"}
readme = "README.md"
dynamic = ["version"]

requires-python = ">=3.8"
dependencies = [
    "numpy>=1.16",
    "backends>=1.6.2",
    "backends-matrix>=1.2.10",
    "plum-dispatch>=2.3.0",
    "stheno>=1.3.10",
    "wbml>=0.3.18",
]

[project.optional-dependencies]
dev = [
    "numpy",
    "pytest>=6",
    "pytest-cov",
    "coveralls",
    "pre-commit",
    "IPython",
    "black==23.7.0",
    "ghp-import",
    "wheel",
    "build",
    "tox",
    "jupyter-book",
    "mypy",
    "pyright",
    "torch",
    "tensorflow",
    "tensorflow-probability[tf]",
]

[project.urls]
repository = "https://github.com/wesselb/neuralprocesses"

[tool.hatch.build]
include = ["neuralprocesses*"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "neuralprocesses/_version.py"

# Tests:

[tool.coverage.run]
branch = true
command_line = "-m pytest --verbose test"
source = ["neuralprocesses"]

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

# Formatting tools:

[tool.black]
line-length = 88
target-version = ["py38", "py39"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
src_paths = ["neuralprocesses", "tests"]

[tool.poetry]
name = "NeuralProcesses"
description = "A framework for composing Neural Processes in Python"
authors = ["Wessel Bruinsma"]
license = "MIT"
readme = "README.md"
version = "0.2.6"

[tool.poetry.dependencies]
python = ">=3.10"
numpy = ">=1.16"
backends = ">=1.6.2"
backends-matrix = ">=1.2.10"
plum-dispatch = ">=2.3.0"
stheno = ">=1.3.10"
wbml = ">=0.3.18"
tensorflow = "^2.17"
tensorflow-probability = "^0.24"
torch = "^2.4"
